Restrict eager loading to @OneToOne and @OneToMany relationships
Previously, a JOIN FETCH was also added when dealing with a
relationship that may return multiple lines (@OneToMany, for example).
Since Hibernate doesn't know how many lines it will need to construct
your entities, it fetches a lot of lines and store them in memory,
resulting in a big performance hit. In that case, the following warning
is dislayed: "HHH000104: firstResult/maxResults specified with
collection fetch; applying in memory!". That commit thus restricts JOIN
FETCH to @OneToOne and @ManyToOne relationships.